move users to repository foders
