+++ server/auth/passport.js
@@ -1,7 +1,7 @@
 const passport = require("passport");
 const passportJWT = require("passport-jwt");
 const { ExtractJwt, Strategy: JwtStrategy } = passportJWT;
-const db = require("../services/database/users");
+const db = require("../Controller/services/repository/users");
 
 const config = {};
 config.secretOrKey = process.env.JWT_SECRET || "your_jwt_secret";
